Workboard re-organize modals
