+++ src/components/projects/Workboard/MyBoard/index.tsx
@@ -3,13 +3,13 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCircleLeft } from '@fortawesome/free-regular-svg-icons';
 import { useState } from 'react';
 import Column from './components/Column';
-import CreateColumnModal from './components/CreateColumnModal';
-import DeleteColumnModal from './components/DeleteColumnModal';
+import CreateColumnModal from '../modals/CreateColumnModal';
+import DeleteColumnModal from '../modals/DeleteColumnModal';
 import { ColToDelete } from '../types';
 import useStore from '../Store';
 import Link from 'next/link';
 import Dropdown from './components/Dropdown';
-import DeleteBoardModal from './components/DeleteBoardModal';
+import DeleteBoardModal from '../modals/DeleteBoardModal';
 import TitleInput from './components/TitleInput';
 import { useRetrieveBoards } from '../useRetrieveBoards';
 import { opaqueIconStyle } from '../ui/tailwindStyles';
